+++ core/src/main/java/org/apache/calcite/rel/rules/ReduceDecimalsRule.java
@@ -317,6 +317,10 @@ public class ReduceDecimalsRule
       real8 = builder.getTypeFactory().createSqlType(SqlTypeName.DOUBLE);
     }
 
+    private RelDataTypeSystem typeSystem() {
+      return builder.getTypeFactory().getTypeSystem();
+    }
+
     /**
      * This defaults to the utility method,
      * {@link RexUtil#requiresDecimalExpansion(RexNode, boolean)} which checks
@@ -345,10 +349,10 @@ public class ReduceDecimalsRule
      * @param scale a scale from one to max precision - 1
      * @return 10^scale as an exact numeric value
      */
+    @SuppressWarnings("deprecation") // [CALCITE-6598]
     protected RexNode makeScaleFactor(int scale) {
       assert scale > 0;
-      assert scale
-          < builder.getTypeFactory().getTypeSystem().getMaxNumericPrecision();
+      assert scale < typeSystem().getMaxNumericPrecision();
       return makeExactLiteral(powerOfTen(scale));
     }
 
@@ -375,20 +379,20 @@ public class ReduceDecimalsRule
      * @param scale a scale from 1 to max precision - 1
      * @return 10^scale / 2 as an exact numeric value
      */
+    @SuppressWarnings("deprecation") // [CALCITE-6598]
     protected RexNode makeRoundFactor(int scale) {
       assert scale > 0;
-      assert scale
-          < builder.getTypeFactory().getTypeSystem().getMaxNumericPrecision();
+      assert scale < typeSystem().getMaxNumericPrecision();
       return makeExactLiteral(powerOfTen(scale) / 2);
     }
 
     /**
      * Calculates a power of ten, as a long value.
      */
+    @SuppressWarnings("deprecation") // [CALCITE-6598]
     protected long powerOfTen(int scale) {
       assert scale >= 0;
-      assert scale
-          < builder.getTypeFactory().getTypeSystem().getMaxNumericPrecision();
+      assert scale < typeSystem().getMaxNumericPrecision();
       return BigInteger.TEN.pow(scale).longValue();
     }
 
@@ -415,10 +419,10 @@ public class ReduceDecimalsRule
      * @param scale a value from zero to max precision - 1
      * @return value * 10^scale as an exact numeric value
      */
-    protected RexNode scaleUp(RexNode value, int scale) {
+    @SuppressWarnings("deprecation") // [CALCITE-6598]
+    protected RexNode scaleUp(SqlParserPos pos, RexNode value, int scale) {
       assert scale >= 0;
-      assert scale
-          < builder.getTypeFactory().getTypeSystem().getMaxNumericPrecision();
+      assert scale < typeSystem().getMaxNumericPrecision();
       if (scale == 0) {
         return value;
       }
@@ -439,9 +443,9 @@ public class ReduceDecimalsRule
      * @return value/10^scale, rounded away from zero and returned as an
      * exact numeric value
      */
-    protected RexNode scaleDown(RexNode value, int scale) {
-      final int maxPrecision =
-          builder.getTypeFactory().getTypeSystem().getMaxNumericPrecision();
+    @SuppressWarnings("deprecation") // [CALCITE-6598]
+    protected RexNode scaleDown(SqlParserPos pos, RexNode value, int scale) {
+      final int maxPrecision = typeSystem().getMaxNumericPrecision();
       assert scale >= 0 && scale <= maxPrecision;
       if (scale == 0) {
         return value;
@@ -484,11 +488,11 @@ public class ReduceDecimalsRule
      * @param scale a value from zero to max precision
      * @return value/10^scale as a double precision value
      */
-    protected RexNode scaleDownDouble(RexNode value, int scale) {
+    @SuppressWarnings("deprecation") // [CALCITE-6598]
+    protected RexNode scaleDownDouble(SqlParserPos pos, RexNode value, int scale) {
       assert scale >= 0;
-      assert scale
-          <= builder.getTypeFactory().getTypeSystem().getMaxNumericPrecision();
-      RexNode cast = ensureType(real8, value);
+      assert scale <= typeSystem().getMaxNumericPrecision();
+      RexNode cast = ensureType(pos, real8, value);
       if (scale == 0) {
         return cast;
       }
@@ -514,9 +518,10 @@ public class ReduceDecimalsRule
      * @return value * 10^scale, returned as an exact or approximate value
      * corresponding to the input value
      */
-    protected RexNode ensureScale(RexNode value, int scale, int required) {
-      final RelDataTypeSystem typeSystem =
-          builder.getTypeFactory().getTypeSystem();
+    @SuppressWarnings("deprecation") // [CALCITE-6598]
+    protected RexNode ensureScale(SqlParserPos pos, RexNode value, int scale,
+        int required) {
+      final RelDataTypeSystem typeSystem = typeSystem();
       final int maxPrecision = typeSystem.getMaxNumericPrecision();
       assert scale <= maxPrecision && required <= maxPrecision;
       assert required >= scale;
@@ -1017,6 +1022,7 @@ public class ReduceDecimalsRule
       super(rexBuilder);
     }
 
+    @SuppressWarnings("deprecation") // [CALCITE-6598]
     @Override public RexNode expand(RexCall call) {
       assert call.getOperator() == SqlStdOperatorTable.FLOOR;
       RexNode decValue = call.operands.get(0);
@@ -1024,11 +1030,12 @@ public class ReduceDecimalsRule
       RexNode value = decodeValue(decValue);
       final RelDataTypeSystem typeSystem =
           builder.getTypeFactory().getTypeSystem();
+      final int maxPrecision = typeSystem.getMaxNumericPrecision();
 
       RexNode rewrite;
       if (scale == 0) {
         rewrite = decValue;
-      } else if (scale == typeSystem.getMaxNumericPrecision()) {
+      } else if (scale == maxPrecision) {
         rewrite =
             makeCase(
                 makeIsNegative(value),
@@ -1067,6 +1074,7 @@ public class ReduceDecimalsRule
       super(rexBuilder);
     }
 
+    @SuppressWarnings("deprecation") // [CALCITE-6598]
     @Override public RexNode expand(RexCall call) {
       assert call.getOperator() == SqlStdOperatorTable.CEIL;
       RexNode decValue = call.operands.get(0);
@@ -1074,11 +1082,12 @@ public class ReduceDecimalsRule
       RexNode value = decodeValue(decValue);
       final RelDataTypeSystem typeSystem =
           builder.getTypeFactory().getTypeSystem();
+      final int maxPrecision = typeSystem.getMaxNumericPrecision();
 
       RexNode rewrite;
       if (scale == 0) {
         rewrite = decValue;
-      } else if (scale == typeSystem.getMaxNumericPrecision()) {
+      } else if (scale == maxPrecision) {
         rewrite =
             makeCase(
                 makeIsPositive(value),