+++ core/src/main/java/org/apache/calcite/rel/type/DelegatingTypeSystem.java
@@ -42,10 +42,16 @@ public class DelegatingTypeSystem implements RelDataTypeSystem {
     return typeSystem.getMaxPrecision(typeName);
   }
 
+  @Override public int getMinPrecision(SqlTypeName typeName) {
+    return typeSystem.getMinPrecision(typeName);
+  }
+
+  @SuppressWarnings("deprecation")
   @Override public int getMaxNumericScale() {
     return typeSystem.getMaxNumericScale();
   }
 
+  @SuppressWarnings("deprecation")
   @Override public int getMaxNumericPrecision() {
     return typeSystem.getMaxNumericPrecision();
   }