+++ core/src/main/java/org/apache/calcite/rel/type/RelDataTypeSystem.java
@@ -54,11 +54,70 @@ public interface RelDataTypeSystem {
    */
   int getMaxPrecision(SqlTypeName typeName);
 
-  /** Returns the maximum scale of a NUMERIC or DECIMAL type. */
-  int getMaxNumericScale();
+  /**
+   * Returns the minimum precision (or length) allowed for this type, or
+   * {@link RelDataType#PRECISION_NOT_SPECIFIED}
+   * if precision/length are not applicable for this type.
+   *
+   * @return Minimum allowed precision
+   */
+  int getMinPrecision(SqlTypeName typeName);
 
-  /** Returns the maximum precision of a NUMERIC or DECIMAL type. */
-  int getMaxNumericPrecision();
+  /** Returns the maximum scale of a NUMERIC or DECIMAL type.
+   * Default value is 19.
+   *
+   * @deprecated Replaced by {@link #getMaxScale}(DECIMAL).
+   *
+   * <p>From Calcite release 1.38 onwards, instead of calling this method, you
+   * should call {@code getMaxScale(DECIMAL)}.
+   *
+   * <p>In Calcite release 1.38, if you wish to change the maximum
+   * scale of {@link SqlTypeName#DECIMAL} values, you should do two things:
+   *
+   * <ul>
+   * <li>Override the {@link #getMaxScale(SqlTypeName)} method,
+   *     changing its behavior for {@code DECIMAL};
+   * <li>Make sure that the implementation of your
+   *     {@code #getMaxNumericScale} method calls
+   *     {@code getMaxScale(DECIMAL)}.
+   * </ul>
+   *
+   * <p>In Calcite release 1.39, Calcite will cease calling this method,
+   * and will remove the override of the method in
+   * {@link RelDataTypeSystemImpl}. You should remove all calls to
+   * and overrides of this method. */
+  @Deprecated // calcite will cease calling in 1.39, and removed before 2.0
+  default int getMaxNumericScale() {
+    return 19;
+  }
+
+  /** Returns the maximum precision of a NUMERIC or DECIMAL type.
+   * Default value is 19.
+   *
+   * @deprecated Replaced by {@link #getMaxScale}(DECIMAL).
+   *
+   * <p>From Calcite release 1.38 onwards, instead of calling this method, you
+   * should call {@code getMaxPrecision(DECIMAL)}.
+   *
+   * <p>In Calcite release 1.38, if you wish to change the maximum
+   * precision of {@link SqlTypeName#DECIMAL} values, you should do two things:
+   *
+   * <ul>
+   * <li>Override the {@link #getMaxPrecision(SqlTypeName)} method,
+   *     changing its behavior for {@code DECIMAL};
+   * <li>Make sure that the implementation of your
+   *     {@code #getMaxNumericPrecision} method calls
+   *     {@code getMaxPrecision(DECIMAL)}.
+   * </ul>
+   *
+   * <p>In Calcite release 1.39, Calcite will cease calling this method,
+   * and will remove the override of the method in
+   * {@link RelDataTypeSystemImpl}. You should remove all calls to
+   * and overrides of this method. */
+  @Deprecated // calcite will cease calling in 1.39, and removed before 2.0
+  default int getMaxNumericPrecision() {
+    return getMaxPrecision(SqlTypeName.DECIMAL);
+  }
 
   /** Returns the LITERAL string for the type, either PREFIX/SUFFIX. */
   @Nullable String getLiteral(SqlTypeName typeName, boolean isPrefix);
@@ -294,22 +353,44 @@ public interface RelDataTypeSystem {
         int s1 = type1.getScale();
         int s2 = type2.getScale();
 
-        final int maxNumericPrecision = getMaxNumericPrecision();
-        int dout =
-            Math.min(
-                p1 - s1 + s2,
-                maxNumericPrecision);
+        final int maxScale = getMaxNumericScale();
+        int six = Math.min(6, maxScale);
+        int d = p1 - s1 + s2;
+        int scale = Math.max(six, s1 + p2 + 1);
+        scale = Math.min(scale, maxScale);
+        int precision = d + scale;
 
-        int scale = Math.max(6, s1 + p2 + 1);
-        scale =
-            Math.min(
-                scale,
-                maxNumericPrecision - dout);
-        scale = Math.min(scale, getMaxNumericScale());
+  // Rules from
+  // https://learn.microsoft.com/en-us/sql/t-sql/data-types/precision-scale-and-length-transact-sql
+        // Rules reproduced here in case the web page goes away.
+        // Note that in SQL server getMaxNumericPrecision() == 38 and
+        // getMaxNumericScale() > 6.
+        // In multiplication and division operations, we need precision - scale places to store
+        // the integral part of the result. The scale might be reduced using the following rules:
+        //
+        // - The resulting scale is reduced to min(scale, 38 - (precision-scale))
+        //   if the integral part is less than 32, because it can't be greater than
+        //   38 - (precision-scale). The result might be rounded in this case.
+        // - The scale isn't changed if it's less than 6 and if the integral part
+        //   is greater than 32. In this case, an overflow error might be raised if
+        //   it can't fit into decimal(38, scale).
+        // - The scale is set to 6 if it's greater than 6 and if the integral part
+        //   is greater than 32. In this case, both the integral part and scale would be
+        //   reduced and resulting type is decimal(38, 6). The result might be rounded to
+        //   7 decimal places, or the overflow error is thrown if the integral part
+        //   can't fit into 32 digits.
+        final int maxPrecision = getMaxNumericPrecision();
+        int bound = maxPrecision - six;  // This was '32' in the MS documentation
+        if (precision <= bound) {
+          scale = Math.min(scale, maxPrecision - (precision - scale));
+        } else {
+          // precision > bound
+          scale = Math.min(six, scale);
+        }
 
-        int precision = dout + scale;
-        assert precision <= maxNumericPrecision;
+        precision = Math.min(precision, maxPrecision);
         assert precision > 0;
+        assert scale <= maxScale;
 
         RelDataType ret;
         ret = typeFactory.
@@ -379,9 +460,12 @@ public interface RelDataTypeSystem {
           return type2;
         }
 
-        int scale = Math.min(Math.max(s1, s2), getMaxNumericScale());
+        final int maxScale = getMaxNumericScale();
+        final int maxPrecision = getMaxNumericPrecision();
+
+        final int scale = Math.min(Math.max(s1, s2), maxScale);
         int precision = Math.min(p1 - s1, p2 - s2) + Math.max(s1, s2);
-        precision = Math.min(precision, getMaxNumericPrecision());
+        precision = Math.min(precision, maxPrecision);
         assert precision > 0;
 
         return typeFactory.createSqlType(SqlTypeName.DECIMAL,