+++ core/src/main/java/org/apache/calcite/rel/type/RelDataTypeSystemImpl.java
@@ -22,6 +22,11 @@ import org.apache.calcite.sql.type.SqlTypeName;
 
 import org.checkerframework.checker.nullness.qual.Nullable;
 
+import java.math.RoundingMode;
+
+import static org.apache.calcite.sql.type.SqlTypeName.DEFAULT_INTERVAL_FRACTIONAL_SECOND_PRECISION;
+import static org.apache.calcite.sql.type.SqlTypeName.MIN_INTERVAL_START_PRECISION;
+
 /** Default implementation of
  * {@link org.apache.calcite.rel.type.RelDataTypeSystem},
  * providing parameters from the SQL standard.
@@ -39,6 +44,7 @@ public abstract class RelDataTypeSystemImpl implements RelDataTypeSystem {
   @Override public int getMaxScale(SqlTypeName typeName) {
     switch (typeName) {
     case DECIMAL:
+      // from 1.39, this will be 'return 19;'
       return getMaxNumericScale();
     case INTERVAL_YEAR:
     case INTERVAL_YEAR_MONTH:
@@ -55,7 +61,37 @@ public abstract class RelDataTypeSystemImpl implements RelDataTypeSystem {
     case INTERVAL_SECOND:
       return SqlTypeName.MAX_INTERVAL_FRACTIONAL_SECOND_PRECISION;
     default:
-      return -1;
+      return RelDataType.SCALE_NOT_SPECIFIED;
+    }
+  }
+
+  /**
+   * Returns the minimum scale (or fractional second precision in the case of
+   * intervals) allowed for this type, or {@link RelDataType#SCALE_NOT_SPECIFIED}
+   * if precision/length are not applicable for this type.
+   *
+   * @return Minimum allowed scale
+   */
+  @Override public int getMinScale(SqlTypeName typeName) {
+    switch (typeName) {
+    case DECIMAL:
+      return 0;
+    case INTERVAL_YEAR:
+    case INTERVAL_YEAR_MONTH:
+    case INTERVAL_MONTH:
+    case INTERVAL_DAY:
+    case INTERVAL_DAY_HOUR:
+    case INTERVAL_DAY_MINUTE:
+    case INTERVAL_DAY_SECOND:
+    case INTERVAL_HOUR:
+    case INTERVAL_HOUR_MINUTE:
+    case INTERVAL_HOUR_SECOND:
+    case INTERVAL_MINUTE:
+    case INTERVAL_MINUTE_SECOND:
+    case INTERVAL_SECOND:
+      return 0; // MIN_INTERVAL_FRACTIONAL_SECOND_PRECISION;
+    default:
+      return RelDataType.SCALE_NOT_SPECIFIED;
     }
   }
 
@@ -69,6 +105,7 @@ public abstract class RelDataTypeSystemImpl implements RelDataTypeSystem {
     case VARBINARY:
       return RelDataType.PRECISION_NOT_SPECIFIED;
     case DECIMAL:
+      // from 1.39, this will be 'return getMaxPrecision(typeName);'
       return getMaxNumericPrecision();
     case INTERVAL_YEAR:
     case INTERVAL_YEAR_MONTH:
@@ -119,6 +156,7 @@ public abstract class RelDataTypeSystemImpl implements RelDataTypeSystem {
   @Override public int getMaxPrecision(SqlTypeName typeName) {
     switch (typeName) {
     case DECIMAL:
+      // from 1.39, this will be 'return 19;'
       return getMaxNumericPrecision();
     case VARCHAR:
     case CHAR:
@@ -152,10 +190,52 @@ public abstract class RelDataTypeSystemImpl implements RelDataTypeSystem {
     }
   }
 
+  /**
+   * Returns the minimum precision (or length) allowed for this type,
+   * or {@link RelDataType#PRECISION_NOT_SPECIFIED}
+   * if precision/length are not applicable for this type.
+   *
+   * @return Minimum allowed precision
+   */
+  @Override public int getMinPrecision(SqlTypeName typeName) {
+    switch (typeName) {
+    case DECIMAL:
+    case VARCHAR:
+    case CHAR:
+    case VARBINARY:
+    case BINARY:
+    case TIME:
+    case TIME_WITH_LOCAL_TIME_ZONE:
+    case TIME_TZ:
+    case TIMESTAMP:
+    case TIMESTAMP_WITH_LOCAL_TIME_ZONE:
+    case TIMESTAMP_TZ:
+      return 1;
+    case INTERVAL_YEAR:
+    case INTERVAL_YEAR_MONTH:
+    case INTERVAL_MONTH:
+    case INTERVAL_DAY:
+    case INTERVAL_DAY_HOUR:
+    case INTERVAL_DAY_MINUTE:
+    case INTERVAL_DAY_SECOND:
+    case INTERVAL_HOUR:
+    case INTERVAL_HOUR_MINUTE:
+    case INTERVAL_HOUR_SECOND:
+    case INTERVAL_MINUTE:
+    case INTERVAL_MINUTE_SECOND:
+    case INTERVAL_SECOND:
+      return MIN_INTERVAL_START_PRECISION;
+    default:
+      return RelDataType.PRECISION_NOT_SPECIFIED;
+    }
+  }
+
+  @SuppressWarnings("deprecation")
   @Override public int getMaxNumericScale() {
     return 19;
   }
 
+  @SuppressWarnings("deprecation")
   @Override public int getMaxNumericPrecision() {
     return 19;
   }