+++ core/src/main/java/org/apache/calcite/sql/type/SqlTypeFactoryImpl.java
@@ -77,6 +77,7 @@ public class SqlTypeFactoryImpl extends RelDataTypeFactoryImpl {
     return canonize(newType);
   }
 
+  @SuppressWarnings("deprecation") // [CALCITE-6598]
   @Override public RelDataType createSqlType(
       SqlTypeName typeName,
       int precision,
@@ -84,9 +85,24 @@ public class SqlTypeFactoryImpl extends RelDataTypeFactoryImpl {
     assertBasic(typeName);
     assert (precision >= 0)
         || (precision == RelDataType.PRECISION_NOT_SPECIFIED);
-    final int maxPrecision = typeSystem.getMaxPrecision(typeName);
-    if (maxPrecision >= 0 && precision > maxPrecision) {
-      precision = maxPrecision;
+    if (precision != RelDataType.PRECISION_NOT_SPECIFIED) {
+      final int minPrecision = typeSystem.getMinPrecision(typeName);
+      final int maxPrecision = typeSystem.getMaxPrecision(typeName);
+      if (maxPrecision >= 0 && precision > maxPrecision) {
+        precision = maxPrecision;
+      }
+      if (precision < minPrecision) {
+        throw RESOURCE.invalidPrecisionForDecimalType(precision, maxPrecision)
+            .ex();
+      }
+    }
+    if (scale != RelDataType.SCALE_NOT_SPECIFIED) {
+      final int minScale = typeSystem.getMinScale(typeName);
+      final int maxScale = typeSystem.getMaxNumericScale();
+      if (scale < minScale) {
+        throw RESOURCE.invalidScaleForDecimalType(scale, minScale, maxScale)
+            .ex();
+      }
     }
     RelDataType newType =
         new BasicSqlType(typeSystem, typeName, precision, scale);
@@ -249,6 +265,7 @@ public class SqlTypeFactoryImpl extends RelDataTypeFactoryImpl {
         : "use createSqlIntervalType() instead";
   }
 
+  @SuppressWarnings("deprecation") // [CALCITE-6598]
   private @Nullable RelDataType leastRestrictiveSqlType(List<RelDataType> types) {
     RelDataType resultType = null;
     int nullCount = 0;