+++ site/_docs/history.md
@@ -48,6 +48,28 @@ using JDK/OpenJDK versions 8 to 19;
 Guava versions 21.0 to 32.1.3-jre;
 other software versions as specified in gradle.properties.
 
+* [a <href="https://issues.apache.org/jira/browse/CALCITE-2067">]
+  **RexLiteral cannot represent accurately floating point values,
+  including NaN, Infinity**.  This fix changes the way RexLiteral
+  represents floating point values.  Previously floating point values
+  were encoded into BigDecimal values.  This caused precision loss
+  when representing the results of simplifying expressions whose
+  results are floating point.  With this change RexLiteral uses
+  internally a Java Double value to represent a SQL DOUBLE, FLOAT, or
+  REAL value.  The result of RexLiteral.getValue() accordingly changes
+  type in this case.
+
+*Deprecated methods in `interface RelDataTypeSystem`*.
+[<a href="https://issues.apache.org/jira/browse/CALCITE-6598">CALCITE-6598</a>]
+deprecates methods `getMaxNumericScale()` and `getMaxNumericPrecision()`,
+to be consistent with `getMinScale(SqlTypeName)` added in
+[<a href="https://issues.apache.org/jira/browse/CALCITE-6560">CALCITE-6560</a>].
+From 1.38, you should instead call `getMaxScale(DECIMAL)` and
+`getMaxPrecision(DECIMAL)`. If you have overridden these methods, Calcite will
+continue to call your overriding method in 1.38 but will cease in 1.39.
+To avoid this *breaking change in 1.39*, during 1.38 you should move your
+override logic to the `getMaxScale` and `getMaxPrecision` methods.
+
 #### New features
 {: #new-features-1-38-0}